machine: do not judge whether a container is booted when terminate
Our boot detection code currently do not work with QEMU well, and
sending SIGRTMIN+4 to an emulated systemd fails too.

As non-booted containers have no dbus to conenct to (which leads to
failure when trying to run something inside it), try to run poweroff
command first, and then running command returns dbus error, assume the
container is not booted (or booted but in some buggy state), and kill
it.

Signed-off-by: Icenowy Zheng <icenowy@aosc.io>