machine: prevent systemd from adjusting terminal title
systemd tries to be annoying by setting a fAnCy terminal title with a
emoji :sparkles: status indicator in it when running command inside a
nspawn container, and adding a tint to the terminal background (which
we suppressed earlier).

Stop, just stop please. Such thing does not help at all. All these bold
moves ...