revert: back to v3.5.2 (v3.9.0)
Unfortunately, the many changes that the commit messages and UI did not
expose to plain sight amounted to a major-version-worthy amount of
changes.

We ran into issues with repository configurations and, without sufficient
documentation about the API and interface changes, the 3.8.x code is
exceedingly difficult to maintain (@eatradish pointed to the
`add/remove-{}` parameter abstraction) and therefore to issue bugfixes.

I have personally communicated (with an apology) to @xtexChooser regarding
this revert and am hoping to re-orient the 3.5.2 - 3.8.8 changes for our
upcoming 4.0 release (with batches of feature work split into pull
requests and have them undergo code and feature reviews).

The version was bumped to v3.9.0 to workaround autopr, which would bump
`app-devel/ciel` if we simply downgraded.